requirements.txt
praw>=7.7.0
beautifulsoup4>=4.12.2
feedparser>=6.0.10
Unidecode>=1.3.6
pygubu>=0.31
PyYAML>=6.0
python-dateutil>=2.8.2
